Update setup_win.cmd; choose 32/64 bit mingw.
With the release of 2.1, it appears somewhere along the chain of building in github actions that the 32-bit isn't successfully building a 32-bit output. In order to ease verifying as well as the added benefit of providing choice, updated the setup_win.cmd to prompt for whether you want 64-bit or 32-bit mingw.